+++ breadbox/breadbox/service/metadata.py
@@ -5,7 +5,12 @@ from breadbox.crud import dataset as dataset_crud
 from breadbox.crud import types as types_crud
 from breadbox.db.session import SessionWithUser
 from breadbox.schemas.custom_http_exception import ResourceNotFoundError
-from breadbox.models.dataset import MatrixDataset, TabularDataset
+from breadbox.models.dataset import (
+    DatasetFeature,
+    DatasetSample,
+    MatrixDataset,
+    TabularDataset,
+)
 from breadbox.service import metadata as metadata_service
 
 from depmap_compute.slice import SliceQuery
@@ -154,3 +159,51 @@ def get_labels_for_slice_type(
         raise ResourceNotFoundError(
             f"Unknown identifier type: '{slice_query.identifier_type}'"
         )
+
+
+def get_dataset_feature_by_label(
+    db: SessionWithUser, dataset_id: str, feature_label: str
+) -> DatasetFeature:
+    """Load the dataset feature corresponding to the given dataset ID and feature label"""
+
+    dataset = dataset_crud.get_dataset(db, db.user, dataset_id)
+    if dataset is None:
+        raise ResourceNotFoundError(f"Dataset '{dataset_id}' not found.")
+    assert isinstance(dataset, MatrixDataset)
+
+    labels_by_given_id = metadata_service.get_dataset_feature_labels_by_id(
+        db, db.user, dataset
+    )
+    given_ids_by_label = {label: id for id, label in labels_by_given_id.items()}
+    feature_given_id = given_ids_by_label.get(feature_label)
+    if feature_given_id is None:
+        raise ResourceNotFoundError(
+            f"Feature label '{feature_label}' not found in dataset '{dataset_id}'."
+        )
+
+    return dataset_crud.get_dataset_feature_by_given_id(
+        db, dataset_id, feature_given_id
+    )
+
+
+def get_dataset_sample_by_label(
+    db: SessionWithUser, dataset_id: str, sample_label: str
+) -> DatasetSample:
+    """Load the dataset sample corresponding to the given dataset ID and sample label"""
+
+    dataset = dataset_crud.get_dataset(db, db.user, dataset_id)
+    if dataset is None:
+        raise ResourceNotFoundError(f"Dataset '{dataset_id}' not found.")
+    assert isinstance(dataset, MatrixDataset)
+
+    labels_by_given_id = metadata_service.get_dataset_sample_labels_by_id(
+        db, db.user, dataset
+    )
+    given_ids_by_label = {label: id for id, label in labels_by_given_id.items()}
+    sample_given_id = given_ids_by_label.get(sample_label)
+    if sample_given_id is None:
+        raise ResourceNotFoundError(
+            f"Sample label '{sample_label}' not found in dataset '{dataset_id}'."
+        )
+
+    return dataset_crud.get_dataset_sample_by_given_id(db, dataset_id, sample_given_id)
